+++ src/w25q128jv.c
@@ -84,7 +84,7 @@ void programPage(uint32_t addr, uint8_t* tx, uint8_t* rx,  uint8_t len) {
     CS_high(GPIOA, 4);
 }
 
-void SPIRead(uint32_t addr, uint8_t* tx, uint8_t* rx, uint8_t len) {
+void SPIRead(uint32_t addr, uint8_t* tx, uint8_t* rx, uint16_t len) {
     while(busy()) {}
     CS_low(GPIOA, 4);
